pizza_app/requirements.txt
asgiref==3.3.1
attrs==20.3.0
coverage==7.3.2
Django==3.1.4
djangorestframework==3.12.2
iniconfig==1.1.1
factory-boy==3.3.0
Faker==19.6.2
freezegun==1.2.2
model-bakery==1.2.1
packaging==20.8
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.1
pytest-cov==4.1.0
pytest-django==4.5.2
pytz==2020.4
sqlparse==0.4.1
toml==0.10.2
requests==2.25.1
